oauth-server/requirements.txt
fastapi
uvicorn[standard]
authlib
python-jose[cryptography]
python-multipart
httpx
jinja2
